+++ docs/changelog.md
@@ -4,7 +4,11 @@
 
 - changed directory structure, added docs, src, and tests
 - created changelog, included CuTest's readme in the docs
-- updates tasks.json in .vscode to configure code generation
+- updated tasks.json in .vscode to configure code generation
 - output file is now `fileopen.exe` due to interpretation of video instructions
 - added .gitignore so we can exclude executables from github
+- Added the testing suite CuTest. More info [here](https://cutest.sourceforge.net/)
+- Added the functions `fileExists` and `filenameHasExtension`
+- Added unit tests for `fileExists` and `filenameHasExtension`
+